Initialize git repository
In order to preserve the changes made over time in terms of this project,
I will be breaking up each change made into its own commit. So note that
while it might seem like I made all of these changes in 1 day in the git
history of this project, in reality I had made these changes over the
course of a few weeks prior to the creation of this repository.