Add starting text descriptions of game
As the player starts in an enclosed tube at the start of the game,
I updated the initial room name and description in order to
better match that scenario.

Instead of making the title of the starting room
"Inside Teleportation Station" or something like that, I decided
to leave it as "???" in order to add mystery at the start,
to make the player question what is going on.