Add ability to examine capsule glass
In making this game, I am designing it around the philosophy that
the player should actively engage with the world around them
in order to progress- learning not only about the world's lore,
but also what to do with it (objectives and things of that nature).

In order to encourage this from the very start, I made the
player's knowledge of the world a small as possible-
being stuck inside of a tube. By creating the glass to examine,
I hope to encourage the player to explore and interact with
as many things as they can when playing the game.