Add jsconfig file to indicate project to VSCode
Prior to adding the jsconfig file, Visual Studio Code
was treating all of the Javascript files of my project as individual
files separate from each other. By adding this config file, it will
treat all of the files as being part of one project. This is especially
useful for implementing hovers properly (the info you get when crtl
hovering over an element in code).