Implement smash command to game.
One thing that I wanted to do with this game is
make it so that the player has to smash their way out of the
capsule that they are in at the start. In order to make this
happen in the context of QuestJS, I had to create a smash command.

Under the glass' smash function, a new temporary exit is created so that
the player can move to the other room in the moment of smashing the glass
and not prior (if I had done it normally, the player could leave the room
earlier without breaking the glass as it would be on the compass rose).