Implement hands equip system
To add some immersion to the game, I wanted to create a system
where the player has to equip certain items in order to use it. For
example, if the player wanted to use a gun, they would have
equip it in their hand(s) before using it (it wouldn't make sense
for the player to be able to shoot a gun if it is in their backpack
right?)

I implemented a basic UI system to display the items in the player's
hands- this will be updated in the near future.

Note that I could have used the RPG Library for weapons and such, but I
wanted to try to implement it on my own in order to better learn
Javascript. This might backfire on me later lol.