Update message system to account for equip hand
After initially creating the equip system, one problem that I quickly
ran into was dealing with the equipMsg string for one handed weapons.
As the player can equip the weapon in either hand, the equipMsg
for that weapon would have to be updated in order to match the
correct hand the player chose.

In order to account for this variablity in hand selection, I made
the equipMsg and unequipMsg strings be templates where the string
"[hands}" gets replaced with the correct hand name based on what
the player chose/the remaining hand that is left.

Note that for two handed weapons this system isn't used as the
only hand string used is "both". Do not use "[hands]" in the string
as it won't get replaced.