Implement proper display GUI for inventory
I included a function that displays the names of the items
that are currently in the player's hands. This consists of
situations where there is 1 item in both hands, 2 separate items
in each hand, 1 item in a hand, and no items in either hands.

As the name being displayed is just a string (the alias of the
item in question), it can be easily updated if needed.