Implement Fallout 3 terminal style appearance
I have begun to customize the appearance of my game in an attempt
to give it its own identity. I decided to follow a similiar appearance
to the terminals that are seen for the hacking minigame in Fallout 3,
with scanlines, green text, and other characteristics. As the game is
being developed, this will be developed alongside it. However, it will
not be a big priority at the moment (as there needs to be a game first
for these changes to the game's appearance to have any effect.